% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateChebyshevCoefficients.R
\name{calculateChebyshevCoefficients}
\alias{calculateChebyshevCoefficients}
\title{Computes the Chebyshev coefficients from a given function and cheb list}
\usage{
calculateChebyshevCoefficients(f, cheb, tolerance = 1e-12)
}
\arguments{
\item{f}{Function to be approximated (function)}

\item{cheb}{List of item from initalizeChebyshevApproximator (list)}

\item{tolerance}{Numerical Tolerance for rounding down}
}
\value{
A list of Chebyshev coefficients (matrix)
}
\description{
Also checks to ensure the cheb$T matrix is orthogonal
The rounding down to 0 in the beginning is to account for numerical precision and is controlled
by the tolerance parameter
The function f only takes one argument
}
