% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateChebyshevPolynomials.R
\name{calculateChebyshevPolynomials}
\alias{calculateChebyshevPolynomials}
\title{Computes the polynomials for a given degree and vector of values.}
\usage{
calculateChebyshevPolynomials(x, N)
}
\arguments{
\item{x}{Vector of values to compute the polynomials at (numeric)}

\item{N}{Highest Degree of the Polynomial (Integer)}
}
\value{
A matrix of the polynomials (matrix)
}
\description{
Resultant matrix of polynomials is of size length(x) by N + 1
}
