% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateChebyshev.R
\name{evaluateChebyshev}
\alias{evaluateChebyshev}
\title{Evaluates the Chebyshev Approximation for a matrix (or a vector) of points}
\usage{
evaluateChebyshev(x, cheb, parallel = FALSE, numcores = 1L)
}
\arguments{
\item{x}{Points to evaluate with size Points by Dimensions (matrix)}

\item{cheb}{List of item from initalizeChebyshevApproximator (list)}

\item{parallel}{Boolean flag for parallelization (logical)}

\item{numcores}{Cores for parallelization (integer)}
}
\value{
A vector of predictions for each point of x
}
\description{
Option for parallelized evaluation for many points to evaluate
}
