% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateChebyshev_T.R
\name{evaluateChebyshev_T}
\alias{evaluateChebyshev_T}
\title{Evaluates the Chebyshev Approximation for a matrix (or a vector) of points and returns
the underlying basis function values instead of the interpolation values}
\usage{
evaluateChebyshev_T(x, cheb, parallel = FALSE, numcores = 1L)
}
\arguments{
\item{x}{Points to evaluate with size Points by Dimensions (matrix)}

\item{cheb}{List of item from initalizeChebyshevApproximator (list)}

\item{parallel}{Boolean flag for parallelization (logical)}

\item{numcores}{Cores for parallelization (integer)}
}
\value{
A matrix of the underlying basis function values
}
\description{
Option for parallelized evaluation for many points to evaluate
}
